XQuartz build script
This adds patches, config options, and the build script to create a signed
XQuartz package and dmg.

Right now, you need to setup some prerequisites manually, but instructions
are provided for doing this using MacPorts.

We are not building some leacy bincompat projects and do not build i386
slices.  As such, this drops some level of binary compatability for legacy
clients.

Signed-off-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>