Build xephyr from src/xorg/xserver
Quite a while ago (XQuartz-2.5?), we started shipping the previously built version of Xephyr
because of issues even when rebuilding Xephyr from the same sources againt newer libraries.

I no longer remember the reason why this was the case, and building Xephyr seems fine, so
let's ship the newer build.

Fixes: https://github.com/XQuartz/XQuartz/issues/33
Signed-off-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>